Tehtävä 6.10 User model admin

+++ app/controllers/breweries_controller.rb
@@ -56,21 +56,20 @@ class BreweriesController < ApplicationController
 
   # DELETE /breweries/1 or /breweries/1.json
   def destroy
-    @brewery.destroy
-    puts "*** PANIMO TUHOTTU ***"
-    kaljat = Beer.all
-    kaljat.each{ |olut|
-      if olut.brewery.nil?
-        puts "panimoton olut: #{olut}"
-        olut.destroy
-      else
-        puts "olut: #{olut.brewery}"
-      end
-    }
+    if current_user.admin == true
+      @brewery.destroy
+      puts "*** PANIMO TUHOTTU ***"
+      Beer.all.each{ |olut| olut.destroy if olut.brewery.nil? }
 
-    respond_to do |format|
-      format.html { redirect_to breweries_url, notice: "Brewery was successfully destroyed." }
-      format.json { head :no_content }
+      respond_to do |format|
+        format.html { redirect_to breweries_url, notice: "Brewery was successfully destroyed." }
+        format.json { head :no_content }
+      end
+    else
+      respond_to do |format|
+        format.html { redirect_to breweries_url, notice: "You do not have admin rights." }
+        format.json { head :no_content }
+      end
     end
   end
